// interfaces/ListUserResponse.ts
// Generated by https://quicktype.io
export interface ListUserResponse {
  data: Data;
  message: string;
  success: boolean;
}

export interface Data {
  page: number;
  per_page: number;
  pre_page: null;
  next_page: null;
  total: number;
  total_pages: number;
  data: IUser[];
}

export interface DataPagination {
  page: number;
  per_page: number;
  pre_page: null;
  next_page: null;
  total: number;
  total_pages: number;
}

export interface IUser {
  _id?: string;
  first_name: string;
  last_name: string;
  document_number: string;
  from_date: string;
  to_date: string;
  from_hour: string;
  to_hour: string;
  createdAt?: string;
  updatedAt?: string;
  __v?: number;
  phone_number?: string;
  state?: boolean;
}

export interface User {
  _id?: string;
  firstName: string;
  lastName: string;
  documentNumber: string;
  fromDate: string;
  toDate: string;
  fromHour: string;
  toHour: string;
  createdAt?: string;
  updatedAt?: string;
  __v?: number;
  phoneNumber?: string;
  state?: boolean;
}